refactor(MVVM-MVI 분할작업중): UI 1차 완료(코드 정리 및 preview, testcode 진행 예정)
