--- presentation/src/main/java/com/android/sj/presentation/interfaces/ui/ScreenNavAirQualityHandler.kt
@@ -1,7 +1,0 @@
-package com.android.sj.presentation.interfaces.ui
-
-interface ScreenNavAirQualityHandler {
-    fun fetchAll(x: String, y: String)
-    fun onStationError(x: String, y: String)
-    fun onAirQualityError()
-}
