--- presentation/src/main/java/com/android/sj/presentation/interfaces/ui/ScreenNavWeatherHandler.kt
@@ -1,8 +1,0 @@
-package com.android.sj.presentation.interfaces.ui
-
-interface ScreenNavWeatherHandler {
-    fun fetchAll(nx: String, ny: String, address: String)
-    fun onNowError(lat: String, lon: String)
-    fun onTimeError(lat: String, lon: String)
-    fun onWeekError(address: String)
-}
