+++ presentation/src/main/java/com/android/sj/presentation/ui/compose/weather/nowweather/WeatherDetailsColumn.kt
@@ -1,0 +1,54 @@
+package com.android.sj.presentation.ui.compose.weather.nowweather
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.dimensionResource
+import com.android.sj.presentation.R
+import com.android.sj.presentation.utils.sp
+
+@Composable
+fun WeatherDetailsColumn(
+    modifier: Modifier,
+    rain: String,
+    wet: String,
+    wind: String
+) {
+    val nowWetRainWindSize = dimensionResource(R.dimen.NowWetRainWindColumn).sp()
+
+    Column(
+        modifier = modifier,
+        horizontalAlignment = Alignment.Start,
+        verticalArrangement = Arrangement.Center
+    ) {
+        Text(
+            text = wet,
+            fontSize = nowWetRainWindSize,
+            modifier = Modifier
+                .weight(1f)
+                .fillMaxWidth()
+                .wrapContentHeight(Alignment.CenterVertically)
+        )
+        Text(
+            text = rain,
+            fontSize = nowWetRainWindSize,
+            modifier = Modifier
+                .weight(1f)
+                .fillMaxWidth()
+                .wrapContentHeight(Alignment.CenterVertically)
+        )
+        Text(
+            text = wind,
+            fontSize = nowWetRainWindSize,
+            modifier = Modifier
+                .weight(1f)
+                .fillMaxWidth()
+                .wrapContentHeight(Alignment.CenterVertically)
+        )
+    }
+}
