+++ presentation/src/main/java/com/android/sj/presentation/ui/compose/weather/weekweather/WeekWeatherItem.kt
@@ -1,15 +1,13 @@
-package com.android.sj.presentation.ui.compose.weather
+package com.android.sj.presentation.ui.compose.weather.weekweather
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
@@ -21,54 +19,17 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import com.android.sj.presentation.R
 import com.android.sj.presentation.models.uimodels.weather.WeekRainSkyUIModel
-import com.android.sj.presentation.state.viewstate.WeatherViewState
-import com.android.sj.presentation.ui.compose.common.UiStateHandler
 import com.android.sj.presentation.ui.previewParam.WeekCardDataPreviewParamProvider
 import com.android.sj.presentation.ui.theme.Color_eceff1
 import com.android.sj.presentation.utils.sp
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
-
-@Composable
-fun WeekWeatherColumn(
-    modifier: Modifier,
-    weatherState: WeatherViewState,
-    errorFunc: () -> Unit
-) {
-    UiStateHandler(modifier, weatherState.weekRainSkyUiState, errorFunc = {errorFunc()}) { successState ->
-        Column(
-            modifier = modifier
-        ) {
-            Text(
-                text = stringResource(R.string.weekWeather),
-                fontSize = dimensionResource(R.dimen.WeatherViewTitle).sp(),
-                fontWeight = FontWeight.Bold,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .align(Alignment.CenterHorizontally)
-            )
-
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(vertical = 8.dp),
-                verticalArrangement = Arrangement.spacedBy(3.dp)
-            ) {
-                items(successState.items.size) { item ->
-                    if(successState.items[item].skyAm != null && successState.items[item].skyPm != null)
-                        WeekWeatherItem(successState.items[item])
-                }
-            }
-        }
-    }
-}
 
 @OptIn(ExperimentalGlideComposeApi::class)
 @Composable
