+++ presentation/src/mvi/java/com/android/sj/presentation/ui/compose/ScreenNav.kt
@@ -6,8 +6,6 @@ import androidx.compose.runtime.remember
 import com.android.sj.presentation.intent.AirQualityIntent
 import com.android.sj.presentation.intent.NaverMapIntent
 import com.android.sj.presentation.intent.WeatherIntent
-import com.android.sj.presentation.interfaces.ui.ScreenNavAirQualityHandler
-import com.android.sj.presentation.interfaces.ui.ScreenNavWeatherHandler
 import com.android.sj.presentation.ui.compose.nav.ScreenNavCommon
 import com.android.sj.presentation.utils.LocalAirQualityVM
 import com.android.sj.presentation.utils.LocalNaverMapVM
@@ -20,47 +18,6 @@ fun ScreenNav(paddingValues: PaddingValues) {
     val weatherVM = LocalWeatherVM.current
     val airQualityVM = LocalAirQualityVM.current
 
-    val weatherHandler = object : ScreenNavWeatherHandler {
-        override fun fetchAll(nx: String, ny: String, address: String) {
-            weatherVM.sendIntent(
-                WeatherIntent.LoadAllWeather(nx = nx, ny = ny, address = address)
-            )
-        }
-        override fun onNowError(lat: String, lon: String) {
-            weatherVM.sendIntent(
-                WeatherIntent.LoadWeather(lat, lon )
-            )
-        }
-        override fun onTimeError(lat: String, lon: String) {
-            weatherVM.sendIntent(
-                WeatherIntent.LoadTimeWeather(lat, lon)
-            )
-        }
-        override fun onWeekError(address: String) {
-            weatherVM.sendIntent(
-                WeatherIntent.LoadWeekRainSky(address)
-            )
-        }
-    }
-
-    val airQualityHandler = object : ScreenNavAirQualityHandler {
-        override fun fetchAll(x: String, y: String) {
-            airQualityVM.sendIntent(
-                AirQualityIntent.LoadAllAirQuality(regionX = x, regionY = y)
-            )
-        }
-        override fun onStationError(x: String, y: String) {
-            airQualityVM.sendIntent(
-                AirQualityIntent.LoadStationFind(x, y)
-            )
-        }
-        override fun onAirQualityError() {
-            airQualityVM.sendIntent(
-                AirQualityIntent.LoadAirQuality
-            )
-        }
-    }
-
     ScreenNavCommon(
         paddingValues = paddingValues,
         mergedEventEffect = remember(naverMapVM, weatherVM, airQualityVM) {
@@ -69,7 +26,15 @@ fun ScreenNav(paddingValues: PaddingValues) {
         getLocation = {
             naverMapVM.sendIntent(NaverMapIntent.GetLocation)
         },
-        weatherHandler,
-        airQualityHandler
+        weatherFetchAll = { nx , ny, address ->
+            weatherVM.sendIntent(
+                WeatherIntent.LoadAllWeather(nx = nx, ny = ny, address = address)
+            )
+        },
+        airQualityFetchAll = { x, y ->
+            airQualityVM.sendIntent(
+                AirQualityIntent.LoadAllAirQuality(regionX = x, regionY = y)
+            )
+        }
     )
 }
